Replace KeyID with additional info
The KeyID is useful in some cases (e.g for OIDC keys), but is not the only
useful information about the key. Expand this into a KeyInfo type, that
includes the ID, the ARN, and the alias if it was loaded that way.

This is added to the Decrypter as well for consistency.